Refactor response serializer in order to conditionally use custom responses
